refactor: 将 app/core/event.py 拆分为 app/core/event 包
按职责拆分单文件 event.py 为包，对外公共 API 与导入时机零改动
（40+ 处 from app.core.event import eventmanager / Event / EventManager 不变）：

- models.py  : 事件值对象 Event + 4 个队列/优先级调优常量（轻量数据层）
- manager.py : EventManager（订阅/广播/链式调度）+ 全局单例 eventmanager
- __init__.py: 急切再导出 Event / EventManager / eventmanager / 常量

设计取舍：
- EventManager 为单个强状态耦合类（私有态贯穿全方法 + 大量 __ 名称改写），
  跨文件拆分会触发名称改写错位而破坏，故整类保留在 manager.py，仅剥离 Event 值对象；
- __init__ 采用急切再导出而非懒加载：历史单文件在 import 期即构造全局单例 eventmanager，
  且无「需独立轻量 import 子模块」的不变量（与 app.core.module 的 loader 不同），
  急切再导出保持导入行为与单例构造时机与历史完全一致；
- __get_class_instance 的 globals() 全局处理器查找迁入 manager.py 后行为不变
  （event.py 顶层 import 均为工具类、无真实处理器类，实际总走动态导入回退）。

全量测试 1713 passed / 19 skipped。
